+++ apps/angular-lib-dev/src/app/datepicker/datepicker.component.ts
@@ -43,7 +43,7 @@ export class DatepickerComponent {
   save() {
     console.log(this.validationForm.value)
   }
-  test($event: any) {
+  test($event: Event) {
     console.log($event)
   }
 }