+++ libs/angular/src/lib/datepicker/datepicker.stories.ts
@@ -25,6 +25,7 @@ export default {
     docs: {
       page: Documentation,
     },
+    componentIds: ['component-datepicker'],
   },
   // More on argTypes: https://storybook.js.org/docs/angular/api/argtypes
   argTypes: {