+++ libs/angular/src/lib/dropdown/dropdown.spec.ts
@@ -17,9 +17,9 @@ describe('Dropdown', () => {
   beforeEach(async () => {
     props = {
       options: [
-        { key: 'A', value: 1 },
-        { key: 'B', value: 2 },
-        { key: 'C', value: 3 },
+        { label: 'A', value: 1 },
+        { label: 'B', value: 2 },
+        { label: 'C', value: 3 },
       ],
     }
     component = await render(NggDropdownComponent, {
@@ -66,18 +66,18 @@ describe('Dropdown', () => {
     })
     it('sets class active on listbox', async () => {
       // initial
-      expect(listbox.className).toEqual('popover popover-dropdown')
+      expect(listbox.className).toEqual('_popover popover-dropdown')
 
       // open
       fireEvent.click(toggleButton)
       await waitFor(() =>
-        expect(listbox.className).toEqual('popover popover-dropdown active')
+        expect(listbox.className).toEqual('_popover popover-dropdown active')
       )
 
       // close
       fireEvent.click(toggleButton)
       await waitFor(() =>
-        expect(listbox.className).toEqual('popover popover-dropdown')
+        expect(listbox.className).toEqual('_popover popover-dropdown')
       )
     })
   })
@@ -98,7 +98,7 @@ describe('Dropdown', () => {
           expect(toggleButton.getAttribute('aria-expanded')).toEqual('false')
         )
         await waitFor(() =>
-          expect(listbox.className).toEqual('popover popover-dropdown')
+          expect(listbox.className).toEqual('_popover popover-dropdown')
         )
       })
       it('sets toggler text', async () => {