+++ libs/extract/src/lib/dropdown/reducers.spec.ts
@@ -5,9 +5,11 @@ import {
   create,
   open,
   select,
+  selectByValue,
   toggle,
   loop,
   highlight,
+  search,
 } from './reducers'
 import { AbstractDropdown, DropdownOption, DropdownTexts } from './types'
 
@@ -23,10 +25,11 @@ describe('dropdown/reducers', () => {
       select: 'Selected',
       selected: 'selected',
       placeholder: 'Select',
+      searchPlaceholder: 'Search',
     }
     options = [
-      { key: 'A', value: 1 },
-      { key: 'B', value: 2 },
+      { label: 'A', value: 1 },
+      { label: 'B', value: 2 },
     ]
   })
   describe('create', () => {
@@ -36,17 +39,17 @@ describe('dropdown/reducers', () => {
     })
     it('uses passed in text', () => {
       const dropdown = create({ id, options, texts })
-      expect(dropdown.texts).toEqual({ ...texts, select: '' })
+      expect(dropdown.texts).toEqual({ ...texts, select: texts.placeholder })
     })
     it('uses default option text', () => {
       const defaultOption = {
-        key: 'C',
+        label: 'C',
         value: 3,
         selected: true,
       }
       const _options = [...options, defaultOption]
       const dropdown = create({ id, options: _options, texts })
-      expect(dropdown.texts.select).toEqual(defaultOption.key)
+      expect(dropdown.texts.select).toEqual(defaultOption.label)
     })
     it('creates an id if not supplied', () => {
       const dropdown = create({ options })
@@ -126,7 +129,7 @@ describe('dropdown/reducers', () => {
         const dropdown = create({ id, options })
 
         expect(dropdown.elements.listbox.classes).toEqual([
-          'popover',
+          '_popover',
           'popover-dropdown',
         ])
       })
@@ -232,11 +235,12 @@ describe('dropdown/reducers', () => {
           undefined,
           true,
         ])
+        expect(dropdown.value).toEqual(2)
       })
       it('sets text', () => {
         const selectedOption = dropdown.options[1]
         dropdown = select(dropdown, selectedOption)
-        expect(dropdown.texts.select).toEqual(selectedOption.key)
+        expect(dropdown.texts.select).toEqual(selectedOption.label)
       })
       it('sets activedecendant', () => {
         const selectedOption = dropdown.options[1]
@@ -252,6 +256,18 @@ describe('dropdown/reducers', () => {
         expect(dropdown.isOpen).toBe(true)
       })
     })
+    describe('selectByValue', () => {
+      it('sets option selected', () => {
+        dropdown = selectByValue(dropdown, dropdown.options[1].value)
+        expect(dropdown.value).toEqual(2)
+      })
+      it('sets empty value to reset previous selection', () => {
+        dropdown = selectByValue(dropdown, dropdown.options[1].value)
+        dropdown = selectByValue(dropdown, undefined)
+        expect(dropdown.value).toEqual(undefined)
+        expect(dropdown.texts.select).toEqual(texts.placeholder)
+      })
+    })
     describe('select multiple', () => {
       beforeEach(() => {
         dropdown = create({ id, options, multiSelect: true })
@@ -263,6 +279,7 @@ describe('dropdown/reducers', () => {
         const selectedOptions = dropdown.options.map((o) => o.selected)
 
         expect(selectedOptions).toEqual([true, true])
+        expect(JSON.stringify(dropdown.value)).toEqual(JSON.stringify([1, 2]))
       })
       it('invert selected option', () => {
         // select
@@ -339,6 +356,38 @@ describe('dropdown/reducers', () => {
             expect(dropdown.options[1].active).toBe(true)
           })
         })
+      })
+    })
+    describe('search', () => {
+      it('adds hidden class from options not containing search text', () => {
+        dropdown = search(dropdown, 'A')
+
+        expect(dropdown.options[0].classes).not.toContain('hidden')
+        expect(dropdown.options[1].classes).toContain('hidden')
+      })
+      it('removes hidden class from options containing search text', () => {
+        dropdown = search(dropdown, 'A')
+        dropdown = search(dropdown, 'B')
+
+        expect(dropdown.options[0].classes).toContain('hidden')
+        expect(dropdown.options[1].classes).not.toContain('hidden')
+      })
+      it('removes hidden class from options containing custom search filter', () => {
+        dropdown.searchFilter = (search, value) => {
+          return value.toString().includes(search)
+        }
+        dropdown = search(dropdown, '1')
+        dropdown = search(dropdown, '2')
+
+        expect(dropdown.options[0].classes).toContain('hidden')
+        expect(dropdown.options[1].classes).not.toContain('hidden')
+      })
+      it('removes hidden class from all options when search text is empty', () => {
+        dropdown = search(dropdown, 'A')
+        dropdown = search(dropdown, '')
+
+        expect(dropdown.options[0].classes).not.toContain('hidden')
+        expect(dropdown.options[1].classes).not.toContain('hidden')
       })
     })
   })