+++ libs/extract/src/lib/dropdown/reducers.ts
@@ -1,13 +1,25 @@
-import { dropdownValues, optionValues } from './defaultValues'
+import { IValidator, validateClassName } from '../helperFunction'
 import { randomId } from '../id'
 import reduce from '../reduce'
+import { dropdownValues, optionValues } from './defaultValues'
 import {
   AbstractDropdown,
   DropdownArgs,
   DropdownOption,
-  ExtendedDropdownOption,
+  DropdownOptionElement,
+  DropdownTexts,
 } from './types'
-import {validateClassName} from "../helperFunction";
+
+const selectionText = (
+  selectedOptions: DropdownOptionElement[],
+  display: string,
+  texts?: DropdownTexts
+) => {
+  const displayValues = selectedOptions.map((option) => option[display])
+  return displayValues?.length > 2
+    ? `${displayValues.length} ${texts?.selected} `
+    : displayValues?.join(', ') || (texts?.placeholder ?? 'Select')
+}
 
 const distinct = <T>(arr: T[]): T[] => {
   const map: Record<string, boolean> = {}
@@ -33,8 +45,8 @@ const extendOption = (
   id: string,
   option: DropdownOption,
   ix: number
-): ExtendedDropdownOption =>
-  reduce<ExtendedDropdownOption>(optionValues, option, {
+): DropdownOptionElement =>
+  reduce<DropdownOptionElement>(optionValues, option, {
     attributes: {
       id: `${id}_option${ix}`,
       role: 'option',
@@ -45,45 +57,36 @@ const extendOption = (
 const extendOptions = (
   options: DropdownOption[],
   id: string
-): ExtendedDropdownOption[] =>
+): DropdownOptionElement[] =>
   options?.map((option, ix) => extendOption(id, option, ix))
 
 export const create = ({
   id = randomId(),
   texts,
   options: _options,
-  useValue,
-  display,
-  selectValue,
+  useValue = 'value',
+  display = 'label',
   loop,
   value,
   multiSelect,
-  validator
+  searchable,
+  validator,
+  compareWith = (o1, o2) => o1 === o2,
+  searchFilter,
 }: DropdownArgs): AbstractDropdown => {
-  useValue = useValue || 'value'
-  display = display || 'key'
-
-  if (value) {
-    _options = _options?.map((option) => ({
-      ...option,
-      selected: option[<string>useValue] === value,
-    }))
-  }
-  const options = extendOptions(_options, id)
-  const selected = options
-    ?.filter((option) => option.selected)
-    .map((option) => option[<string>display])
-  const dropdown: Partial<AbstractDropdown> = {
+  const options = extendOptions(_options, id) ?? []
+  const selectedOptions = options.filter((o) => o.selected)
+  const selectedValues = selectedOptions.map((option) => option[useValue])
+  let dropdown: Partial<AbstractDropdown> = {
     id,
+    value: multiSelect ? selectedValues : selectedValues[0],
     texts: {
       close: texts?.close ?? 'Close',
       optionsDescription: texts?.optionsDescription ?? 'Options',
       placeholder: texts?.placeholder ?? 'Select',
+      searchPlaceholder: texts?.searchPlaceholder ?? 'Search',
       selected: texts?.selected ?? 'selected',
-      select:
-        selected?.length > 2
-          ? `${selected.length} ${texts?.selected} `
-          : selected?.join(', '),
+      select: selectionText(selectedOptions, display, texts),
     },
     elements: {
       toggler: {
@@ -99,10 +102,17 @@ export const create = ({
     options,
     isLooping: loop,
     isMultiSelect: multiSelect,
+    isSearchable: searchable,
     useValue,
     display,
-    selectValue,
-    validator
+    validator,
+    compareWith,
+    searchFilter,
+  }
+
+  // Override intital selected property options if value is provided
+  if (value) {
+    dropdown = selectByValue(dropdown as AbstractDropdown, value)
   }
 
   return reduce(dropdown, dropdownValues)
@@ -122,6 +132,7 @@ export const open = (dropdown: AbstractDropdown): AbstractDropdown =>
       },
     },
   })
+
 export const close = (dropdown: AbstractDropdown): AbstractDropdown => {
   if (dropdown.isOpen) {
     return reduce(dropdown, {
@@ -146,20 +157,33 @@ export const toggle = (dropdown: AbstractDropdown): AbstractDropdown => {
   const newDD = dropdown.isOpen ? close(dropdown) : open(dropdown)
   return newDD
 }
+
 export const highlight = (
   dropdown: AbstractDropdown,
-  selection: ExtendedDropdownOption | number
+  selection: DropdownOptionElement | number
 ): AbstractDropdown => {
-  let option: ExtendedDropdownOption
+  let option: DropdownOptionElement
   if (typeof selection === 'number') {
     const opts = dropdown.options
-    const currentlySelectedIndex = opts.findIndex((o) => o.active)
+
+    const visibleDropdownOptions = opts.filter(
+      (element) => !element.classes.includes('hidden')
+    )
+
+    const currentlySelectedIndex = visibleDropdownOptions.findIndex(
+      (o) => o.active
+    )
     let newSelectedIndex = currentlySelectedIndex + selection
+
     if (newSelectedIndex < 0)
-      newSelectedIndex = dropdown.isLooping ? opts.length - 1 : 0
-    if (newSelectedIndex >= opts.length)
-      newSelectedIndex = dropdown.isLooping ? 0 : opts.length - 1
-    option = opts[newSelectedIndex]
+      newSelectedIndex = dropdown.isLooping
+        ? visibleDropdownOptions.length - 1
+        : 0
+    if (newSelectedIndex >= visibleDropdownOptions.length)
+      newSelectedIndex = dropdown.isLooping
+        ? 0
+        : visibleDropdownOptions.length - 1
+    option = visibleDropdownOptions[newSelectedIndex]
   } else {
     option = selection
   }
@@ -175,7 +199,10 @@ export const highlight = (
       reduce(o, {
         active: option === o,
         attributes: {},
-        classes: option === o ? ['active', 'sg-highlighted'] : [],
+        classes:
+          option === o
+            ? ['active', 'sg-highlighted']
+            : o.classes.filter((obj) => obj !== 'active' && 'sg-highlighted'),
       })
     ),
   })
@@ -183,21 +210,56 @@ export const highlight = (
 
 export const select = (
   dropdown: AbstractDropdown,
-  selection: ExtendedDropdownOption
+  selection: DropdownOptionElement
 ): AbstractDropdown => {
-  const isSelected = (option: ExtendedDropdownOption) => {
-    const isTarget = selection[dropdown.useValue] === option[dropdown.useValue]
+  const selectedValue = selection[dropdown.useValue]
+  const currentValue = dropdown.value
+  let value
 
-    if (dropdown.isMultiSelect) {
-      //Invert the selected option and keep others like previously
-      return isTarget ? !option.selected : option.selected
+  if (dropdown.isMultiSelect && Array.isArray(currentValue)) {
+    const index = currentValue.findIndex((v: any) =>
+      dropdown.compareWith(v, selectedValue)
+    )
+
+    // Toggle selected value
+    if (index < 0) {
+      value = [...currentValue, selectedValue]
     } else {
-      return isTarget
+      value = [
+        ...currentValue.slice(0, index),
+        ...currentValue.slice(index + 1),
+      ]
+    }
+  } else {
+    value = dropdown.isMultiSelect ? [selectedValue] : selectedValue
+  }
+
+  return selectByValue(dropdown, value)
+}
+
+export const selectByValue = (
+  dropdown: AbstractDropdown,
+  selection?: any
+): AbstractDropdown => {
+  const isSelected = (value1: any) => {
+    if (!selection) {
+      return false
+    } else if (dropdown.isMultiSelect && Array.isArray(selection)) {
+      return selection.some((value2) => dropdown.compareWith(value1, value2))
+    } else {
+      return dropdown.compareWith(selection, value1)
     }
   }
 
+  if (selection && dropdown.isMultiSelect && !Array.isArray(selection)) {
+    console.warn(
+      'Dropdown is marked as multiselect but recieved a non-array value:',
+      selection
+    )
+  }
+
   const options = dropdown.options.map((option) => {
-    const selected = isSelected(option)
+    const selected = isSelected(option[dropdown.useValue])
     return reduce(option, {
       selected,
       attributes: {
@@ -205,16 +267,12 @@ export const select = (
       },
     })
   })
-
   const selectedOptions = options.filter((o) => o.selected)
-  const displayValues = selectedOptions.map((o) => o[dropdown.display])
+
   return reduce(dropdown, {
+    value: selection,
     texts: {
-      select:
-        displayValues?.length > 2
-          ? `${displayValues.length} ${dropdown.texts?.selected} `
-          : displayValues?.join(', ') ||
-            (dropdown.texts?.placeholder ?? 'Select'),
+      select: selectionText(selectedOptions, dropdown.display, dropdown.texts),
     },
     elements: {
       listbox: {
@@ -235,27 +293,31 @@ export const blur = (dropdown: AbstractDropdown): AbstractDropdown =>
   reduce(dropdown, {
     isTouched: dropdown.isTouched || !dropdown.isOpen,
   } as Partial<AbstractDropdown>)
+
 export const active = (
   dropdown: AbstractDropdown,
   isActive: boolean
 ): AbstractDropdown =>
   reduce(dropdown, { isActive } as Partial<AbstractDropdown>)
+
 export const loop = (
   dropdown: AbstractDropdown,
   isLooping: boolean
 ): AbstractDropdown =>
   reduce(dropdown, { isLooping } as Partial<AbstractDropdown>)
+
 export const keypress = (
   dropdown: AbstractDropdown,
   key: string,
   event: KeyboardEvent
 ): AbstractDropdown | undefined => {
-  const opts = dropdown.options
+  const opts = dropdown.options.filter(
+    (option) => !option.classes.includes('hidden')
+  )
+
   let action
+
   switch (key) {
-    //case ' ':
-    //  action = toggle(dropdown)
-    //  break
     case 'Escape':
       if (dropdown.isOpen) action = close(dropdown)
       break
@@ -281,8 +343,18 @@ export const keypress = (
       action = open(highlight(dropdown, opts[opts.length - 1]))
       break
     case ' ':
-    case 'Enter':
-      // eslint-disable-next-line no-case-declarations
+    case 'Enter': {
+      if (dropdown.isSearchable && dropdown.isOpen && key === ' ') {
+        const searchFieldIsActive =
+          document.activeElement ===
+          document.getElementById(
+            dropdown.elements.toggler?.attributes?.id + '_search-input'
+          )
+        if (searchFieldIsActive) {
+          break
+        }
+      }
+
       const activeOption = opts.find((option) => option.active)
       action =
         dropdown.isOpen && activeOption
@@ -293,6 +365,7 @@ export const keypress = (
             : close(select(dropdown, activeOption))
           : toggle(dropdown)
       break
+    }
     default:
       break
   }
@@ -309,12 +382,34 @@ export const popper = (
 ): AbstractDropdown =>
   reduce(dropdown, { elements: { listbox: { attributes: { style } } } })
 
-export const validate = (dropdown: AbstractDropdown, validator: any) => reduce(
-  dropdown, {
+export const validate = (dropdown: AbstractDropdown, validator: IValidator) =>
+  reduce(dropdown, {
     elements: {
       toggler: {
-        classes: addClass(dropdown.elements?.toggler?.classes, validateClassName(validator?.indicator))
-      }
-    }
-  }
-)
+        classes: addClass(
+          dropdown.elements?.toggler?.classes,
+          validateClassName(validator?.indicator)
+        ),
+      },
+    },
+  })
+
+export const search = (
+  dropdown: AbstractDropdown,
+  searchInput: string
+): AbstractDropdown =>
+  reduce(dropdown, {
+    options: dropdown.options.map((option) => {
+      const searchIncludes = (value: any): boolean =>
+        value?.toString().toLowerCase().includes(searchInput.toLowerCase())
+
+      const isMatch =
+        searchInput.length === 0 ||
+        searchIncludes(option[dropdown.display]) ||
+        dropdown.searchFilter?.(searchInput, option[dropdown.useValue])
+
+      return isMatch
+        ? { ...option, classes: removeClass(option.classes, 'hidden') }
+        : { ...option, classes: addClass(option.classes, 'hidden') }
+    }),
+  } as Partial<AbstractDropdown>)