+++ libs/react/src/lib/dropdown/dropdown.spec.tsx
@@ -1,7 +1,6 @@
 import { act, render, RenderResult, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { Dropdown, DropdownProps } from './dropdown'
-import button from "../form/button/button";
 
 const tick = (t = 0) => new Promise<void>((r) => setTimeout(r, t))
 jest.setTimeout(50000)
@@ -15,9 +14,9 @@ describe('Dropdown', () => {
   beforeEach(async () => {
     props = {
       options: [
-        { key: 'A', value: 1 },
-        { key: 'B', value: 2 },
-        { key: 'C', value: 3 },
+        { label: 'A', value: 1 },
+        { label: 'B', value: 2 },
+        { label: 'C', value: 3 },
       ],
     }
     component = render(<Dropdown {...props} />)
@@ -65,18 +64,18 @@ describe('Dropdown', () => {
       const user = userEvent.setup()
 
       // initial
-      expect(listbox.className).toEqual('popover popover-dropdown')
+      expect(listbox.className).toEqual('_popover popover-dropdown')
 
       // open
       await user.click(toggleButton)
       await waitFor(() =>
-        expect(listbox.className).toEqual('popover popover-dropdown active')
+        expect(listbox.className).toEqual('_popover popover-dropdown active')
       )
 
       // close
       await user.click(toggleButton)
       await waitFor(() =>
-        expect(listbox.className).toEqual('popover popover-dropdown')
+        expect(listbox.className).toEqual('_popover popover-dropdown')
       )
     })
   })
@@ -100,7 +99,7 @@ describe('Dropdown', () => {
           expect(toggleButton.getAttribute('aria-expanded')).toEqual('false')
         )
         await waitFor(() =>
-          expect(listbox.className).toEqual('popover popover-dropdown')
+          expect(listbox.className).toEqual('_popover popover-dropdown')
         )
       })
       it('sets toggler text', async () => {