+++ libs/chlorophyll/CHANGELOG.md
@@ -2,6 +2,91 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.0.0-beta.36](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.35...@sebgroup/chlorophyll@1.0.0-beta.36) (2022-11-11)
+
+
+
+# [1.0.0-beta.35](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.34...@sebgroup/chlorophyll@1.0.0-beta.35) (2022-11-10)
+
+
+
+# [1.0.0-beta.34](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.33...@sebgroup/chlorophyll@1.0.0-beta.34) (2022-11-08)
+
+
+
+# [1.0.0-beta.33](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.32...@sebgroup/chlorophyll@1.0.0-beta.33) (2022-11-08)
+
+
+
+# [1.0.0-beta.32](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.31...@sebgroup/chlorophyll@1.0.0-beta.32) (2022-11-08)
+
+
+
+# [1.0.0-beta.31](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.30...@sebgroup/chlorophyll@1.0.0-beta.31) (2022-11-08)
+
+
+### Features
+
+* 🎸 Styling for headers ([d5c0597](https://github.com/sebgroup/green/commit/d5c0597b48f878817d7af118bb8b980b00035075))
+
+
+
+# [1.0.0-beta.30](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.29...@sebgroup/chlorophyll@1.0.0-beta.30) (2022-10-26)
+
+
+
+# [1.0.0-beta.29](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.28...@sebgroup/chlorophyll@1.0.0-beta.29) (2022-10-12)
+
+
+
+# [1.0.0-beta.28](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.27...@sebgroup/chlorophyll@1.0.0-beta.28) (2022-10-07)
+
+
+
+# [1.0.0-beta.27](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.26...@sebgroup/chlorophyll@1.0.0-beta.27) (2022-10-03)
+
+
+
+# [1.0.0-beta.26](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.25...@sebgroup/chlorophyll@1.0.0-beta.26) (2022-09-27)
+
+
+
+# [1.0.0-beta.25](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.24...@sebgroup/chlorophyll@1.0.0-beta.25) (2022-09-23)
+
+
+
+# [1.0.0-beta.24](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.23...@sebgroup/chlorophyll@1.0.0-beta.24) (2022-09-20)
+
+
+
+# [1.0.0-beta.23](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.22...@sebgroup/chlorophyll@1.0.0-beta.23) (2022-09-19)
+
+
+### Features
+
+* 🎸 Dropdown - support search filter [#473](https://github.com/sebgroup/green/issues/473) ([4883331](https://github.com/sebgroup/green/commit/48833313c99661d83fa26066c42f12ff93e26d3c))
+
+
+### BREAKING CHANGES
+
+* 🧨 -
+
+
+
+# [1.0.0-beta.23](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.22...@sebgroup/chlorophyll@1.0.0-beta.23) (2022-09-19)
+
+
+### Features
+
+* 🎸 Dropdown - support search filter [#473](https://github.com/sebgroup/green/issues/473) ([4883331](https://github.com/sebgroup/green/commit/48833313c99661d83fa26066c42f12ff93e26d3c))
+
+
+### BREAKING CHANGES
+
+* 🧨 -
+
+
+
 # [1.0.0-beta.22](https://github.com/sebgroup/green/compare/@sebgroup/chlorophyll@1.0.0-beta.21...@sebgroup/chlorophyll@1.0.0-beta.22) (2022-08-31)